setrecognizer/requirements.txt
Pillow==8.1.0
requests==2.25.1
Flask==1.1.2
numpy
opencv-python
pyyaml
imgaug
pyclipper
lmdb
selenium